requirements.txt
notebook
sqlalchemy
psycopg2-binary
python-dotenv
pandas
tqdm
datasets
transformers
ray